// frontend/src/api/models/handlerRechargeRequestPaymentMethod.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerRechargeRequestPaymentMethod =
  (typeof HandlerRechargeRequestPaymentMethod)[keyof typeof HandlerRechargeRequestPaymentMethod]

export const HandlerRechargeRequestPaymentMethod = {
  CASH: 'CASH',
  WECHAT: 'WECHAT',
  ALIPAY: 'ALIPAY',
  BANK: 'BANK',
} as const